refactor: #49. 예외 클래스 패키지 분리
작업한 것
- 예외 핸들러와 BaseException, BadExceptionRequest 클래스 common 패키지 아래로 이동
- 도메인 아래에 exception 패키지 생성
- 유저와 관련한 exception class 들 user.exception 패키지로 이동

앞으로 할 것
- 도메인별로 핸들러 함수 적용
- 적절한 응답 코드를 반환하도록 수정
- 도메인별로 예외 클래스 생성
