+++ dionysos-api/src/main/java/com/dionysos/api/common/exception/BaseException.java
@@ -1,11 +1,15 @@
-package com.dionysos.api.exception;
+package com.dionysos.api.common.exception;
 
 import com.dionysos.api.common.model.ErrorModel;
+import lombok.Builder;
+import lombok.Getter;
 
+@Getter
 public class BaseException extends RuntimeException {
 
-    protected ErrorModel error;
+    private ErrorModel error;
 
+    @Builder
     protected BaseException(ErrorModel error) {
         super(error.getMessage(),
                 null
