+++ dionysos-api/src/main/java/com/dionysos/api/common/response/code/DionysosAPIErrorCode.java
@@ -1,4 +1,4 @@
-package com.dionysos.api.common.utils;
+package com.dionysos.api.common.response.code;
 
 import lombok.Getter;
