+++ dionysos-api/src/main/java/com/dionysos/api/diary/service/DiaryService.java
@@ -5,8 +5,8 @@ import com.dionysos.api.diary.dto.RequestUpdateDiaryDto;
 import com.dionysos.api.diary.dto.ResponseDiaryDto;
 import com.dionysos.api.diary.entity.Diary;
 import com.dionysos.api.diary.repository.DiaryRepository;
-import com.dionysos.api.exception.BadRequestException;
-import com.dionysos.api.exception.UnAuthorizedException;
+import com.dionysos.api.common.exception.BadRequestException;
+import com.dionysos.api.user.exception.UnAuthorizedException;
 import com.dionysos.api.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
